requirements.txt
parliament-functions==0.1.0
python-dotenv==0.21.0
requests==2.28.1
slack-sdk==3.2.1
